server/requirements.txt
# Django API dependencies
Django==4.2.24
setuptools>=65,<70
PyJWT>=2.8,<3
requests>=2.31,<3
python-dotenv>=1,<2 
gunicorn>=21,<24
